requirements.txt
fastapi[standard]
sqlalchemy
asyncpg
alembic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
jinja2
python-dotenv
sqladmin
email-validator
